pyproject.toml
[project]
name = "django-shop-ku4eraviy"
version = "0.1.0"
description = "Онлайн-магазин на Django"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-debug-toolbar (>=6.2.0,<7.0.0)",
    "pillow (>=12.1.1,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "ipython (>=9.10.0,<10.0.0)"
]
